move views to app
